refactor(file-structure): change App and chore
