+++ src/index.tsx
@@ -3,7 +3,7 @@ import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import { store } from "./app/store";
 
-import App from "./App";
+import App from "./app/App";
 
 ReactDOM.render(
   <Provider store={store}>
